build: eliminate all compiler warnings on armv7/aarch64/x86-64
Full three-architecture rebuild (corei7-64, aarch64-halium,
cortexa8t2hf/armv7-neon) now compiles with -Wall -Wextra -Wconversion and
zero warnings, zero errors. Changes:

- inc/core/MojAutoPtr.h: explicitly default/declare the same-type copy
  assignment operators on MojAutoPtr, MojAutoArrayPtr, MojSharedPtr,
  MojSharedArrayPtr and MojSharedFreePtr; their implicit versions were
  flagged deprecated because each class also declares a copy constructor
  or destructor.
- MojDbCursor, MojDbKind, MojDbSearchCursor/MojDbSearchCacheCursor:
  reordered member-initializer lists to match declaration order
  (-Wreorder); no behavior change, initialization order was already
  determined by declaration order regardless of initializer-list order.
- inc/db/MojDbServiceHandler.h, inc/db/MojDbSpaceAlert.h,
  inc/luna/MojLunaService.h, test/luna-service/MojLunaServiceTest.cpp:
  each declares a differently-shaped handleCancel() on a class deriving
  from MojSignalHandler, which hides the base virtual instead of
  overloading it; add `using MojSignalHandler::handleCancel;` so both
  signatures stay callable (-Woverloaded-virtual).
- inc/engine/sandwich/MojDbSandwichDatabase.h: likewise pull in
  MojDbStorageExtDatabase::get/del via using-declarations instead of
  hiding them.
- MojJson.cpp, MojDbQueryPlan.cpp, inc/core/internal/MojNumberLexer.h:
  annotate the switch-case fall-throughs that were already intentional
  (and commented as such) with [[fallthrough]] (-Wimplicit-fallthrough).
- MojDbTextTokenizer.cpp: ubrk_safeClone() is deprecated since ICU 52;
  switch to ubrk_clone(), which also drops the now-unused stack buffer.
- MojDb.cpp, MojDbKind.cpp, MojDbServiceHandler.cpp: brace the bodies of
  four single-statement if/else clauses that gcc's -Wmisleading-indentation
  flagged as ambiguous; verified none of them were an actual logic bug
  (see the audit in the previous commits for the ones that were).
- MojDbServiceHandlerBase.cpp: check the exit status of the fault-register
  script's system() call instead of casting it away (-Wunused-result).
- MojDbShardEngine.cpp, MojDbSandwichEngine.cpp, MojDbSearchCacheCursor.cpp,
  test/*: fix narrowing conversions flagged by -Wconversion with explicit,
  range-appropriate casts (MojInt64->size_t/unsigned long only after a
  non-negative check, iterator distance ->MojUInt32, etc).
- test/dbclient-perf/DbGenerator.cpp: MojSize is 32-bit on armv7 but the
  format string used %lu; switch to %llu with an explicit widening cast so
  the same code is correct on both ILP32 and LP64 targets.
- CMakeLists.txt: mark -Wno-cast-function-type/-Wno-address/
  -Wno-nonnull-compare as accepted (see inline comments: the
  CategoryHandler dispatch tables intentionally cast between
  same-ABI member-function-pointer types, and the -Waddress/
  -Wnonnull-compare hit is a single false positive on a
  reference-to-pointer conversion); mark the third-party
  leveldb-tl include path SYSTEM so its own warnings don't surface here.
- MojDbSandwichDatabase.cpp: while chasing the last -Wmaybe-uninitialized
  warning (leveldb-tl's Part default constructor leaves `prefix`
  genuinely uninitialized, safe here because every caller checks
  useShard()'s error before touching `part`), found and fixed a real bug
  in del(): the transactional-delete branch discarded the leveldb::Status
  returned by part.Delete() instead of assigning it to `st`, so `st` kept
  its default (ok) value and every transactional delete unconditionally
  reported foundOut=true while silently swallowing any real delete error.
  Documented the remaining false positive with a narrow, explained pragma
  rather than patching the third-party header.

Verified with a from-scratch build (rm -rf build dir) on all three
target toolchains via the OE cross sysroots in
/media/herrie/LuneOS/wrynose/webos-ports; ninja reports 0 warnings on
each. Reran the full gtest/gtest-style suite (mojocore, db-core, sandwich,
leveldb, dbclient-perf, db test runner) plus the new stress harness on
x86-64 against the sandwich backend: all pass. gtest_dbclient and the
MojUtil core test both require a live luna-service2 hub socket
(/var/run/luna-service2/com.palm.hub) and segfault without one; confirmed
identical behavior in the pristine pre-existing prebuilt binaries, so this
is a pre-existing test-environment gap (these tests need to run on-device
or against a running ls-hubd), not a regression from this branch.

+++ test/dbclient-perf/DbGenerator.cpp
@@ -40,7 +40,7 @@ MojErr generateDatabase(size_t count, const char * outputFile)
     for (MojSize i = 0; i < count; ++i) {
         MojObject obj;
         MojString entry;
-        entry.format(_T("{\"_id\":%lu,\"_kind\":\"LunaDbClientTest:1\",\"foo\":\"test_foo\",\"bar\":\"test_bar\"}"), i + 1);
+        entry.format(_T("{\"_id\":%llu,\"_kind\":\"LunaDbClientTest:1\",\"foo\":\"test_foo\",\"bar\":\"test_bar\"}"), (unsigned long long)(i + 1));
         err = obj.fromJson(entry);
         MojErrCheck(err);
         err = db.put(obj);
